CommandResult with generics

+++ atm/src/main/java/ru/otus/slisenko/atm_department/core/department/AtmDepartmentImp.java
@@ -22,7 +22,7 @@ public class AtmDepartmentImp implements AtmDepartment {
     @Override
     public int getTotalBalance() {
         return atmSet.stream()
-                .mapToInt((atm) -> atm.notify(new GetBalanceCommand()).getResult())
+                .mapToInt((atm) -> (int) atm.notify(new GetBalanceCommand()).getResult())
                 .sum();
     }
